pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.1.1",
]
